+++ packages/editor-common/web/src/Icons/index.ts
@@ -1,11 +1,6 @@
 // @create-index
 
-export { default as AlignTextCenterIcon } from './AlignTextCenterIcon';
-export { default as AlignJustifyIcon } from './AlignJustifyIcon';
-export { default as AlignLeftIcon } from './AlignLeftIcon';
-export { default as AlignRightIcon } from './AlignRightIcon';
 export { default as DropdownArrowIcon } from './DropdownArrowIcon';
-export { default as LinkIcon } from './LinkIcon';
 export { default as ImageIcon } from './AnchorsThumbnails/ImageIcon';
 export { default as VideoIcon } from './AnchorsThumbnails/VideoIcon';
 export { default as GalleryIcon } from './AnchorsThumbnails/GalleryIcon';