+++ packages/plugin-divider/web/src/toolbar/selectors.js
@@ -5,7 +5,7 @@ import {
   AlignCenterIcon,
 } from 'wix-rich-content-plugin-commons';
 
-import { AlignLeftIcon, AlignRightIcon } from 'wix-rich-content-editor-common';
+import { AlignLeftIcon, AlignRightIcon } from 'wix-rich-content-ui-components';
 
 import {
   SIZE_LARGE,