+++ packages/ui-components/web/src/Icons/index.ts
@@ -31,3 +31,8 @@ export { default as SearchIcon } from './SearchIcon';
 export { default as DropdownArrowIcon } from './DropdownArrowIcon';
 export { default as CheckIcon } from './CheckIcon';
 export { default as ErrorIcon } from './ErrorIcon';
+export { default as AlignTextCenterIcon } from './AlignTextCenterIcon';
+export { default as AlignJustifyIcon } from './AlignJustifyIcon';
+export { default as AlignLeftIcon } from './AlignLeftIcon';
+export { default as AlignRightIcon } from './AlignRightIcon';
+export { default as LinkIcon } from './LinkIcon';